+++ src/Application/Event/Meeting/MeetingsDeletedAllEvent.php
@@ -8,22 +8,33 @@
  * @author Elao <contact@elao.com>
  */
 
-namespace Proximum\Vimeet\Application\Command\Meeting\Admin;
+namespace Proximum\Vimeet\Application\Event\Meeting;
 
 use Proximum\Vimeet\Domain\Model\Admin;
 use Proximum\Vimeet\Domain\Model\Event;
+use Symfony\Component\EventDispatcher;
 
-class DeleteAll
+class MeetingsDeletedAllEvent extends EventDispatcher\Event
 {
     /** @var Event */
-    public $event;
+    private $event;
 
     /** @var Admin */
-    public $admin;
+    private $admin;
 
     public function __construct(Event $event, Admin $admin)
     {
         $this->event = $event;
         $this->admin = $admin;
+    }
+
+    public function getEvent(): Event
+    {
+        return $this->event;
+    }
+
+    public function getAdmin(): Admin
+    {
+        return $this->admin;
     }
 }